moved TopBarSharingButtonsView
